+++ gemeaux/exceptions.py
@@ -7,3 +7,9 @@ class ImproperlyConfigured(Exception):
     """
     When the configuration is a problem
     """
+
+
+class TemplateError(Exception):
+    """
+    When there's an issue with templating
+    """